finished merge sort and made some refactoring

+++ data_structures/include/static_array.h
@@ -1,7 +1,7 @@
 #ifndef DYNAMIC_ARRAY_H
 #define DYNAMIC_ARRAY_H
 
-int* createStaticArray(size_t size);
+int *createStaticArray(int size);
 void destroyStaticArray(int *array);
 
 #endif
